requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.1.0
python-decouple==3.8
psycopg2-binary==2.9.9
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
redis==5.0.1
reportlab==4.0.7
openpyxl==3.1.2
pandas==2.1.4

# Seguridad — Fase 1.2
django-ratelimit==4.1.0
django-axes==6.4.0

# Monitoreo y auditoría — Fase 5
sentry-sdk==2.55.0
django-auditlog==3.4.1

# WebSocket — Fase 7
daphne==4.2.1  # servidor ASGI para WebSocket en desarrollo

# Documentación de API — Fase 8
drf-spectacular==0.29.0
